rrs_lib
Trait Memory

pub trait Memory: Downcast {
    // Required methods
    fn read_mem(&mut self, addr: u32, size: MemAccessSize) -> Option<u32>;
    fn write_mem(
        &mut self,
        addr: u32,
        size: MemAccessSize,
        store_data: u32,
    ) -> bool;
}

A trait for objects which implement memory operations

Required Methods

fn read_mem(&mut self, addr: u32, size: MemAccessSize) -> Option<u32>

Read size bytes from addr.

addr must be aligned to size. Returns None if addr doesn’t exist in this memory.

fn write_mem(&mut self, addr: u32, size: MemAccessSize, store_data: u32) -> bool

Write size bytes of store_data to addr

addr must be aligned to size. Returns true if write succeeds.

Implementations

impl dyn Memory

pub fn is<__T: Memory>(&self) -> bool

Returns true if the trait object wraps an object of type __T.

pub fn downcast<__T: Memory>(self: Box<Self>) -> Result<Box<__T>, Box<Self>>

Returns a boxed object from a boxed trait object if the underlying object is of type __T. Returns the original boxed trait if it isn’t.

pub fn downcast_rc<__T: Memory>(self: Rc<Self>) -> Result<Rc<__T>, Rc<Self>>

Returns an Rc-ed object from an Rc-ed trait object if the underlying object is of type __T. Returns the original Rc-ed trait if it isn’t.

pub fn downcast_ref<__T: Memory>(&self) -> Option<&__T>

Returns a reference to the object within the trait object if it is of type __T, or None if it isn’t.

pub fn downcast_mut<__T: Memory>(&mut self) -> Option<&mut __T>

Returns a mutable reference to the object within the trait object if it is of type __T, or None if it isn’t.